rt: don't set/clear network driver options in rt-tune-guest.cfg
Issue: LINUXPG-1057
Issue: LINCD-8389

The idea of the rt-tune files was to set/unset options that would
reflect favourably on reducing latency.  This was mostly the domain of
"soft" options, but could also include hardware options, at a generic
level - such as disabling USB.

However, the option lines removed here are for enabling/disabling
specific endpoint network drivers.  That is the domain of a BSP and
shouldn't be here.

The only reason these lines are here is because it happens to reflect
the hardware configuration of a ten year old target that was of interest
to a specific customer at the genesis of OVP.

There is a remote chance this might break some automated test but that
would be a good thing, since we would want to revisit what that test is
doing and whether it is of any value anymore.

We also fix an obvious missing hash mark on an "is not set" line.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>