gcc8/9: adapt to recent '-dev' package change in oe-core
Issue: LINCD-9785

oe-core recently changed the '-dev' package to rrecommends
instead of depends on the main package. This makes setting
RDEPENDS:xxx-dev to "" having no effect of breaking the linkage.
So set RRECOMMENDS:xxx-dev to "" to break the linkage.

This patch deliberately keeps the RDEPENDS settings in place
as it does not harm and it ensures we still break the linkage
in case oe-core later wants to switch back to previous behavior.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>