qemu: vcpu: restore CONFIG_KVM "poison" functionality
Issue: LINCD-9511

Restore the CONFIG_KVM poison warning functionality.

CONFIG_KVM should not be used in common code.  If we look at:
commit cbca3722a33c ("include/exec/poison: Mark CONFIG_KVM as
poisoned, too") we see that it states:

  CONFIG_KVM is only defined for target-specific code, so nobody
  should use it by accident in common code. To avoid such subtle
  bugs, CONFIG_KVM is now marked as poisoned in common code.

If we remove the poison, we just make the error silent, which is
definitely Not Good.`

As the poison delete wasn't part of the original author's commit,
this makes the uprev more in keeping with preserving history and
not implicitly growing other people's Signed-off-By: scope.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>