CGL: drop old copy of RT cyclictest
Issue: LINUXEXEC-23264

We have the current version via normal OE content in:

	meta/recipes-rt/rt-tests

This old version dates back to products based on v2.6.x kernels.

At that time, we had a separate CGL kernel type that used a portion of
the RT patches to deliver threaded IRQ support.

Threaded IRQ support has been a part of mainline for a long time now,
and hence this old copy can simply be removed.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>