CGL: centralize all KERNEL_FEATURES into distro features.
Issue: LINUXEXEC-23264

While packages can depend on certain kernel config options, we can't
extrapolate that to assume the package is the only user relying on that
kernel option set.

By grouping all the various kernel config additions together in with the
existing distro kernel features, we ensure we retain all the prior
settings for a future audit.

This also opens the door to treat the package and config audits
relatively separately - so that package audits based on what use cases
need to be preserved can be performed now.

Note that we aren't really concerned with the ability to still isolate
CUT-enabling options from the larger CGL set - at this point the goal is
primarily kernel config data preservation and package/config separation.

We re-format the list on an item-per-line basis, since we expect to be
deleting individual lines when we get to the kernel config audit.  That
will most likely also involve corresponding removals from the
yocto-kernel-cache repository where appropriate.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>