CGL: drop old GPLv2 userspace monit package.
Issue: LINUXEXEC-23264

We see in this older commit why we still have this old GPLv2 version
of this package:

  commit f485184e48ef2935d023401490520db9d8445ce2
  Author: Yi Zhao <yi.zhao@windriver.com>
  Date:   Fri Nov 1 14:13:28 2019 +0800

    wrlinux-cgl-whitelist.inc: remove monit5

    Issue: LIN1019-2614

    Remove monit 5 as it is under AGPLv3 license. Add zabbix to whitelist as
    replacement. We still keep monit 4.9 and install it by default because
    there are some CGL CUT cases depends on it.

Even if we weren't phasing out CGL, we'd still want to get rid of this
for the same reason meta-gplv2 is going away.  Hanging onto old
non-maintained snapshot versions of packages is not ideal from a
security standpoint.

As we've already removed CUT in a prior series, we don't have that
dependency issue to be concerned about either.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>