CGL: drop core of old CUT (CGL Unit Test) suite
Issue: LINUXEXEC-23264

The surrounding packages in this layer provide infrastructure to
support these unit tests - an in-house creation which was created
roughly 15y ago to provide validation of components of the CGL
v5.0 specification.

Rather than pick through the scripts and untangle dependencies on a per
package basis, we are better to just drop the test scripts and harness
all in one at an early stage in the overall scheme of things.

Even if we decide that some test script is worth saving and porting
elsewhere, it doesn't change the fact they can't live here anymore.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>