CGL: remove last remaining file - kernel config/scc references
Issue: LINUXEXEC-23264

This was intentionally left as the last file for removal as it
will probably be the one and only thing we might want to come
back to and reference.

Looking at the list, I'm not seeing anythnig that I can justify
pushing down into an "always present" condition, given that the
CGL use case was a lot more server-oriented, but that is not
true when considering our customers as a whole.

We could, if there was demand for it, have some kind of a simple
feature/server template that added things like iSCSI and nfsd but
I won't recommend doing that right here and now.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>