wrlinux: move PREFERRED_VERSION from most templates
Issue: LINUXEXEC-23239

Now that WRL_RECIPES have been moved out of templates we have enabled
recipes being built without including templates. So we we can't rely
on PREFERRED_VERSION set in layers and need to set these in layer.conf
so they are always applied.

Some PREFERRED_VERSION remain in templates, but this is only when a
template exists to accomplish the goal of using a different version
than would typically be in use.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>