recipelists: add openIPMI stuff from what was wrlinux-cgl
Issue: LINCD-10656

As part of the demand driven restoration of packages that were previously a
part of wrlinux-cgl, this restores openIPMI to the supported recipe list.  Note
also that they are restored at a top level in keeping with the recipe changes
in LINUXEXEC-23239.

The net-snmp appears as it is a dependency of openipmi.

We may also need to add ipmitool, but since I don't have an advanced
understanding of which IPMI packages provide what features, I'll leave that
to be demand-driven as well and only address the specific defect here.

Other packages can be nominated for inclusion in LINUXEXEC-23480

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>