feature/tensorflow: ensure distro features include tensorflow
Issue: LINCD-10947

For now, when tensorflow template is enabled, 3.11 python is built
for native and 3.10 is built for target. In order to ensure all
python versions (native, target, nativesdk) match, we need to set
the distro features correctly.

As a side effect for current incorrect setting, python3-installer-native
build fails. With this patch, the distro features are set correct and
3.10 is used for all python variants, then python3-installer-native
build succeeds.

We deliberately use ':append' instead of '+=' here because the
'tensorflow' distro feature is a must here, it's not optional. To avoid
it being accidently overridden, we use ':append' here.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>