qemu: transfer across VCPU prio patches
Issue: LINUXEXEC-23705

These commits from the discontinued OVP allow qemu to parse

  -vcpu 0,affinity=0x2,prio=0 -vcpu 1,affinity=0x4,prio=80

type arguments in order to better tune the system.  In theory, one can
get the same result with "chrt" and "taskset" but the reality is that
qemu runs each VCPU in a thread, which you can only "see" with "ps -eLf"
and even then you've not got any clear mapping between threads and VCPU.

Add to that, the arg parsing in chrt and taskset are horrible.

So, as much as I'd like to put these patches behind us, I think we
probably are best to continue them for a little while longer.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>