networkmanager: make use of /proc/net/pnp for externally configured eth0
Issue: LINUXEXEC-24057

In case externally configured network interface, NM will not try
to reconfigure it and it can of course not get domain server information
via DHCP protocol. This happens when 'ip=dhcp' is specified in
kernel command line parameter. But we want the /etc/resolv.conf contents
to be usable. So we should make use of /proc/net/pnp[1].

In WRLinux's current network framework, the automatically configured
interface in case of 'ip=dhcp' will always to eth0. So we name the
dispatcher script[1] eth0-ip-cmdline.sh.

[1] https://www.kernel.org/doc/Documentation/admin-guide/nfs/nfsroot.rst
[2] https://man.archlinux.org/man/NetworkManager-dispatcher.8.en

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>