wrlinux-ovp: delete all remaining WRL_RECIPE entries
Issue: LINUXEXEC-23498

Having dealt with the entries with a more obvious destiny, what is left
here is the remainder.  In order to keep our maintenance burden
minimized (one of the goals of this effort) we will assume these are no
longer required and were only for OVP specific use cases.

If needed, we will bring some back on demand under LINUXEXEC-23500.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>