wrlinux: qemu: Enable vhost by default
Issue: LINCD-11420

To fix the following runtime error,

$ virsh start guest1
error: Failed to start domain 'guest1'
error: internal error: qemu unexpectedly closed the monitor:
2023-01-10T03:09:35.238495Z qemu-system-x86_64: -netdev tap,fd=39,id=hostnet0,vhost=on,vhostfd=41:
vhost-net support is not compiled in
2023-01-10T03:09:35.238571Z qemu-system-x86_64: -netdev tap,fd=39,id=hostnet0,vhost=on,vhostfd=41:
warning: vhost-net requested but could not be initialized
qemu-system-x86_64: /usr/src/debug/qemu/7.2.0-r0/net/net.c:1106: net_client_init1: Assertion `nc' failed.

vhost along with a bunch of other PACKAGECONFIG options of qemu was enabled
by defaut for OVP. When we removed OVP, these options were considered not
having clear reason to exist and purposely dropped. They are expected to be
added back in an on-demand and as-needed manner later.

vhost is a widely used feature and necessary to be enabled by default.

Signed-off-by: He Zhe <zhe.he@windriver.com>