conf/layer.conf: remove PREFERRED_VERSION_cni
Issue: LINUXEXEC-23246

The local cni version was added because at that time upstream was
using a non-working dev version. Now upstream is using a stable
version and from the last few upgrade commits, it's unlikely to
start tracking a dev version again. We now start to sync with
upstream as there's little point keeping a local version.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>