feature/preserve-buildtime: set rootfs timestamp to actual task execution time
Issue: LINCD-7143

The REPRODUCIBLE_ROOTFS_TIMESTAMPS will be set anyway in current Yocto's codes,
and yocto has explicitly refused to maintain two paths. However, some users are
confused about the timestamps of some rootfs files.
e.g.
Mar 9 2018 fsl-s32g2xxa-evb.dtb

The 'Mar 9 2018' comes from the default value of REPRODUCIBLE_ROOTFS_TIMESTAMPS,
"1520598896".

To solve this problem, set REPRODUCIBLE_ROOTFS_TIMESTAMPS to current time. As
bitbake does the evaluation of the variable when it is retrieved, this essentially
means that the time is set to the time the corresponding task is executed. And
this is what we expect.

Note that the naming is a little inacurate because after set to current time,
REPRODUCIBLE_ROOTFS_TIMESTAMPS does not have the effect of 'reproducible' any more.
But it would be better than maintaining a prone-to-change local patch, given the
fact that this REPRODUCIBLE_ROOTFS_TIMESTAMPS has been used in three different
bbclasses.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>