wrlinux-installer.conf: skip template feature/initramfs
Issue: LINCD-12354

The installer image uses ${installer_image_name}-initramfs as initramfs
image. But template feature/initramfs overwrites it and use
wrlinux-image-initramfs instead which starts ttyUSB0 when boot the
system. It does not work with distro wrlinux-installer and fails to boot
the installer image, so skip the template feature/initramfs.

Signed-off-by: Kai Kang <kai.kang@windriver.com>