nativesdk-packagegroup-sdk-host: add nativesdk-rpm-build
Issue: LINCD-12278

For old release such as LTS22, we have nativesdk-rpm-build in
SDK by default. This is more of a coincidence than a deliberate
design. The nativesdk-rpm-build was brought into SDK because of
the following dependency chain.

  nativesdk-opkg -> nativesdk-libsolv -> nativesdk-rpm -> nativesdk-rpm-build

The nativesdk-rpm had a dependency on nativesdk-rpm-build because
in old RPM releases before 4.18, nativesdk-rpm requires librpmbuild.so
which was provided by nativesdk-rpm-build, resulting in the runtime
dependency being added automatically by OE.

Now we have users using rpmbuild inside SDK, and we had it available
for previous releases' SDKs, let's explicitly add it to SDK.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>