selinux-setlabels.bbclass: add selinux_setlabels_handler
Issue: LINCD-12417

There are some unlabeled systemd links generated by systemd_preset_all
function in wic image. This is because in IMAGE_PREPROCESS_COMMAND,
systemd_preset_all function runs after selinux_setlabels function, which
causes the files it creates to be unlabeled.

The selinux_setlabels function should run as late as possible. To
guarantee that, we append it to IMAGE_PREPROCESS_COMMAND in
RecipePreFinalise event handler, this ensures it is the last function in
IMAGE_PREPROCESS_COMMAND.

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>