Common native package list update
Issue: LINCD-438

boost-build-native is involved by
   vsomeip -> boost -> boost-build-native
and vsomeip is enabled on an aptiv board before but the board is not supported by CD.

As for the other changes, all except python3-pyelftools-native are imported by python3-cryptography-native.
python3-cryptography-native has been removed from dependencies of optee-os in layer meta-ti

   * bdfc4149 optee-os: Upgrade to upstream 3.18.0

but they still exist in layer meta-arm

   meta-arm/recipes-security/optee/optee-os-3_19.inc
   15:DEPENDS = "python3-pyelftools-native python3-cryptography-native"

The meta-arm layer will not be supported any more in CD so the python3 related native packages are removed.

vim-native is involved by atf-s32g from layer nxp-s32g which is not supported by CD too.

Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>