wrlinux-recipelists.conf: drop unsupported qtdoc, qtpdf
Issue: LINCD-14774

Upstream added qtdoc in packagegroup-qt6-addons.bb but we don't need or want that recipe.
Qtdoc pulls in qtpdf which is essentially chromium in qt clothing.
We decided when we started on qt6 that we'd avoid chromium/qtwebengine unless there
was a customer request since it's a large, difficult to securely maintain package
and we already have chromium without qt's wrapper.

Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>