feature/qt-essentials: add wayland DISTRO_FEATURE
Issue: LINCD-14337

Temporarily add wayland to qt-essentials' list of DISTRO_FEATURES
to unblock the qtwayland configure and build. i.e. avoid the error:
    ERROR: Qt Wayland Client requires QtGui to be build with support for wayland

The qtwayland or some meta-qt6 recipe should be changed using:
   REQUIRED_DISTRO_FEATURES
but I'm not going to take the time now to test this with poky only.
This change builds and one of the qtwayland examples even runs in qemu.

I've opened the following upstream bug:
    https://bugreports.qt.io/browse/QTBUG-123489
so this LINCD bug should remain open.

Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>