ostree: review and handle local patches
Issue: LINUXEXEC-31784

Local patches are reviewed one by one. For those that are specific
to WRLinux, change the status from 'LAT specific' to 'WRLinux specific',
and also add detailed comments about why it's specific to WRLinux.

The 0004-deploy-using-etc-in-runtime-as-merge-source.patch has some
problem. Its contents does not match its title and comments. I can see
that maybe in very early releases of ostree, ostree does not do 3-way
merge of /etc, resulting in a local patch trying to merge /etc instead
of /usr/etc. But for current ostree release, /etc is handled pretty well.
The actual contents of this patch is actually adding a WRLinux specific
environment variable. So change its title and comments and its file name
as well.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>