ostree: remove lat related contents
Issue: LINUXEXEC-31784

The 'no-initramfs' is LAT specific. Remove these contents from
wrlinux. They are moved back to the meta-lat layer and will take
effect only when 'lat' is in DISTRO_FEATURES. The ostree_lat.inc
is renamed to ostree_ostree.inc just as how it was named in previous
wr-ostree layer.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>