systemd: remove unnecessary ostree settings
Issue: LINUXEXEC-31784

Removing 'link-udev-shared' does not give much benefit now, at least
for the current systemd recipe. This is because the shared libraries
are not in the big systemd package, they're in libsystemd-shared and
libsystemd. There's the size difference with/without link-udev-shared
for initramfs-ostree-image:

With link-udev-shared: 31M
Without link-udev-shared: 30M

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>