conf/wrlinux-vendor-revision.conf: Add initial VENDOR_REVISIONs
Issue: LINUXEXEC-32007

All recipes will have a initial VR '.vr2401', the 'vr' prefix is to distinguish
with PR Server.

Use pn_pepv as the key is because there might be multiple versions for a
recipe, and different versions recipes have different patches.

It will be enabled by enable-vendor-revision.bbclass.

The initial size is 696K and is plain text which shouldn't increase much size
for the repo when upgrade.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>