qttools: avoid install llvm cmake in sysroot
Issue: LIN1024-6527

This issue is related to the fix for LIN1024-5855. The qttools_%.bbappend was added
to the wrong folder. when building the Poky distro image, layer qt6 is not available,
the following error show:
"ERROR: No recipes in default available for:
  layers/amd-x86/recipes-graphics/qttools/qttools_%.bbappend"

Move the previous change to the dynamic-layers here to fix the qttools configuration
error on the amd-snowyowl-64 machine.

Both LLVM and Clang are provided in the amd-snowyowl-64 configuration.
Additionally, Mesa requires more LLVM libraries to build, which results
in the installation of unnecessary LLVM CMake files in qttools ${RECIPES_SYSROOT}.

However, qttools only requires Clang to build, and the presence of these
extra LLVM CMake files causes confusion during CMake's detection of LLVM and Clang.

To resolve this issue, using SSTATE_EXCLUDEDEPS_SYSROOT to avoid the unnecessary
LLVM CMake files install to qttools ${RECIPES_SYSROOT} to ensure CMake correctly
detects and uses Clang.

Signed-off-by: Zhang Peng <peng.zhang1.cn@windriver.com>