wrlinux-recipelists.conf: Add python3-gpt-image
Issue: LINUXEXEC-34178

python3-gpt-image is to create GUID Partition Table disk images on local disks.
Written in pure Python gpt-image allows GPT disk images to be built on a local
filesystem and exported to a destination device. This is useful for creating
a disk image on SD Cards or embedded devices.

Signed-off-by: Libo Chen <libo.chen.cn@windriver.com>